Add template parsing to cred class
