+++ lib/list.js
